finished day 11

+++ src/main/java/software/aoc/day11/Runner11.java
@@ -14,7 +14,7 @@ public record Runner11(List<String> input) {
         }
     }
 
-    public int run() {
+    public long run() {
         return PathFinder.from(input).paths();
     }
