+++ src/main/java/software/aoc/day11/a/Main11A.java
@@ -4,7 +4,7 @@ import software.aoc.day11.Runner11;
 
 public class Main11A {
     static void main() {
-        int result = Runner11.with(Main11A.class.getClassLoader().getResourceAsStream("Day11Input.txt"))
+        long result = Runner11.with(Main11A.class.getClassLoader().getResourceAsStream("Day11Input.txt"))
                 .run();
         System.out.println(result);
     }
