+++ src/test/java/Day11ATest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class Day11ATest {
     @Test
     public void test() {
-        int result = Runner11.with(this.getClass().getResourceAsStream("Day11ATestInput.txt"))
+        long result = Runner11.with(this.getClass().getResourceAsStream("Day11ATestInput.txt"))
                 .run();
         assertThat(result).isEqualTo(5);
     }
